Update paths with separate folder for static and JS files
